start moving connection-handler consumer files in advance of factoring out consumer state module
