+++ flo-client-lib/src/codec/mod.rs
@@ -1,7 +1,6 @@
 #[cfg(feature = "serde-json-codec")]
 mod serde;
 
-use std::string::FromUtf8Error;
 use std::error::Error;
 
 use event::OwnedFloEvent;
